bqsr: Improve cigar expansion
Keep track of an alignment window on the read that excludes clipped
regions, along with the corresponding reference window.

Also track another read window which excludes clipped regions as
well as leading and trailing insertion regions (required for base
alignment quality calculations).

Also renames some of the fields in the cigar context.

Change-Id: Idfef966fc214ba986fe30f3ba8282b014d48516e

[git-p4: depot-paths = "//sw/gpgpu/nvbio/bqsr/": change = 18886535]