bqsr: Clean up BAM batch and reference sequence usage
Collect both host and device batches in a common BAM_alignment_batch
structure and always pass this structure to any host functions.

Clean up our usage of the reference sequence to make sure we always get
it from the context. Switch the context to holding both the host and
device sequences (the context view still holds only the device
sequence).

Change-Id: Iab99dfd35a13a6bfe78492f651a80e12c0e53736

[git-p4: depot-paths = "//sw/gpgpu/nvbio/bqsr/": change = 18886552]