Add size information to vector views; allow copying from view to vector
In order to allow using a view to wrap the host-side data, include size
information in both vector and packed_vector views. Also add a method
to copy from a view into a vector/packed_vector.

This is required to allow referencing a shared memory segment directly
in the sequence data view.

[git-p4: depot-paths = "//sw/gpgpu/nvbio/bqsr/": change = 18886595]