covariates_table: Move sort implementation into common header
Moves the CUB sorter from covariates_table to a generic implementation
in primitives/parallel.h and modifies it so that we pass the actual
number of bits in the key to the sort routine.

Also includes some cosmetic fixes on the bit constants in the bit
packer structure to make this usage clear.