baq: Add lmem variant
Adds a variant of the HMM implementation that runs all steps in a
single large kernel, using local memory to store the forward and
backward matrices. This yields a ~2x speed-up in BAQ, but requires a
fixed maximum read length.

The old code is kept in case we want to implement a hybrid
version later on.