snp_filter: Rework SNP filtering
Refactor the SNP filtering to iterate over SNP ranges and not make use
of the read offset list, but rather use the cigar expansion data
instead. Move cigar expansion to the top of the pipeline.

Fix the SNP filtering algorithm itself to properly detect overlaps of
SNP ranges with reads.