Load missing reference sequences while processing
Adds a generation counter to keep track of changes to the reference
sequence during processing. If the I/O thread encounters an alignment
that refers to a sequence which was not yet loaded, the sequence is
loaded and the generation counter is updated.

Batches track the reference sequence generation when they're loaded.
The worker thread updates the device version of the reference if a
change is detected.