snp_filter: Rework the search algorithm
The previous search algorithm did not account for SNPs that are larger
than a sequencing read, but it turns out there are in fact many such
SNPs.

Replace the sorted index for feature end positions with a prefix scan
of the end positions using the max operator. This acts as an index that
maps a database position to the largest end position of any SNP to the
left of that index in the database.

Use the new search index when searching overlaps by doing a lower bound
search using the read start position as the key. This replaces most of
the merging code from before with an extra binary search.