snp_filter: Do not move clipped deletions to the left
If a feature starts on a deleted base, GATK extends the covered
interval to the nearest M/I base on the left. However, if a feature
lies on a deleted base outside the current clipping window, that
feature is ignored.

Do not extend coverage window to the left if the anchor is outside the
clipping window.