cigar: Fix failure detection in getReadCoordinateForReferenceCoordinate
If a deletion occurs starting at the second base of a read and the
first base is part of the sequencing adapter,
getReadCoordinateForReferenceCoordinate would get confused and failed
to find the right answer.

Check for failure based on the read coordinate and not on the event
coordinate. When iterating backwards over the read, pre-decrement and
check for wrap-around.