baq: Fix reference and query window calculations
Match the GATK reference window query calculation, where the reference
window start position is the base alignment for the read minus the
number of leading insertions minus half of the minimum band width (and
similarly for the end position).

The query window is always equal to the clipped reference window.