baq: Do not cap BAQ qualities for "odd" reads
When iterating through the CIGAR operators, if the end position doesn't
match the read length, GATK considers the read to be "odd" and
overwrites the BAQ output with the original quality values.

Implement the same behavior in cap_baq_reads: iterate over the read
without updating qualities to determine "oddness"; if read is not odd,
iterate again and update qualities.

Note that we should probably modify this to compute "oddness" early on,
so as to avoid computing BAQ for "odd" reads.