baq: Store HMM reference window in read coordinates
cap_baq_qualities requires the HMM reference window in read
coordinates. However, because the window is expected to start to the
left of the base alignment point in the reference, doing the backwards
transform can be problematic.

Switch to read-local coordinates for the HMM reference window. This
simplifies the code somewhat and also reduces the memory footprint of
the HMM reference window array.