Handle reads aligned to chromosomes not referenced in dbSNP
If a read aligns to a chromosome that does not appear in the dbSNP, the
reference database and resident map will contain more slots than the
dbSNP database.

Modify segmented_database to properly handle the case where the
resident segment map is larger than the storage. When running SNP
filtering, check if the chromosome exists in the dbSNP.