Reinstate serialization+mmap support
Adds back the reference/dbsnp serialization support and supporting code
to read/write to system shared memory locations.

The approach is a little simpler than before as we no longer care about
minimizing memory usage (this is meant solely as a debug tool to speed
up the process of bisecting large input files).