baq: Clamp HMM window start position
When a read aligns to the start of a chromosome, the resulting HMM
window start position will be negative. Clamp HMM window position to 0
and adjust reference length accordingly.

This fixes a long-standing bug with reads aligned to the start of a
chromosome. Note that a similar bug has been present in GATK for a long
time, results might diverge between GATK and Firepony for this case.