baq: More HMM window clamping fixes
Instead of reducing the HMM window length in the reference by the same
amount that was truncated, always subtract 1 instead. This mimics the
same (erroneous?) behavior implemented in GATK.

Compute the expected shift in read base pair locations that is encoded
in the BAQ state. Apply that shift to the expected position when
transforming the quality scores.