baq: Allow BAQ to disable reads when HMM window falls off the reference
GATK skips processing reads for which the end of the computed HMM
window lies beyond the end of the reference sequence.

Make compute_hmm_windows disable the read and emit a special token
value when the HMM window is detected to be invalid. Remove disabled
reads from the active read lists by compaction.