firepony: Automatic batch size selection
Pick a suitable batch size based on the list of enabled GPUs. The batch size
across all GPUs is constrained based on the GPU with the smallest amount of
memory.

Batch sizes are empirically determined based on Illumina reads and assuming a
~50% error rate. Note that this is not yet guaranteed to always work, but
should remove the need to manually tune the batch size for most common cases.

As a side effect, this change causes Firepony to refuse to run on GPUs with
less than 4GB of memory due to performance concerns with very small batch
sizes.